Generated by Spark: I want you to create an admin panel with authentication, from where I can add or modify the selection options in each data layer with option to upload each raster tif file or point shapefile and its classification. So that I can directly upload the rasters and shapefiles and their associated classification and you can maintain the directory and naming convention through code. If its a shapefile, make sure to have an option to select from a dropdown which has names of all its attributes and we can select the name of the attribute which corresponds to design capacity based on which the size of point shapefile can be decided. Also have an option to select the icon to display for point shapefile, also have an option to upload a png file that can be used as point shapefile icon to be overlaid on map instead of simple circles.
If its a raster, show the min max and mean value of raster, and give options to set the 5 classes making sure the values are relevant i.e. not less than min and not greater than max, that is the first class should start with min and last class should end with max and also ask for colour hex code for each class. Also, have an option like same as with the dropdown of past uploaded rasters or shapefile for any new upload, so that if the classification or legend is same as any past uploaded raster or shapefile, then we dont need to reselect again. have this for each section indiviually like classification, design capacity variable selection, hex codes for each class, because sometimes we mighthave that the classification is different but colour for each class is same. In that case we can set new classification and select same colour hex codes as past upload selections.
Also ask to upload boundary of each country and ask the attribute name which we want to display on map on hover over the region. As I told you the boundary will be adm1 level so it will have boundary of each province, so when we hover over each province, it should shows its name.